% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R
\name{bi.degree}
\alias{bi.degree}
\title{bi.degree}
\usage{
bi.degree(filename, filetype = 0)
}
\arguments{
\item{filename}{input file name}

\item{filetype}{0}
}
\description{
Get the degree list
}
\examples{
degreelist = bi.degree("example1.el")

}

