% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R
\name{bi.format}
\alias{bi.format}
\title{Add number of vertices and edges to the input file}
\usage{
bi.format(filename, filetype = 0)
}
\arguments{
\item{filename}{Input file name}

\item{filetype}{Input file format <default = 0>. 0-edge list, 1-binary matrix.}
}
\description{
This funtion will calculate the number of vertices and edges and add them to the head of the input file.
The original input file will be changed.
}
\examples{
\dontrun{
bi.format("example2.el")
bi.format("example5.bmat", 1)
}

}

