% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R
\name{bi.print}
\alias{bi.print}
\title{bi.print}
\usage{
bi.print(bi, text = 0)
}
\arguments{
\item{bi}{bicliques}

\item{text}{default is 0. If you want to show labels, change it to 1}
}
\description{
print the bicliques
}
\examples{
bi.print(biclique)

}

