% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R, R/package-biclique.R
\docType{package}
\name{biclique}
\alias{biclique}
\alias{biclique-package}
\title{biclique}
\usage{
biclique(filename)
}
\arguments{
\item{filename}{input file name}
}
\description{
Biclique enumeration

biclique package.
}
\examples{
biclique("example1.el")

}
\author{
Yuping Lu
}
\keyword{biclique}

