% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R
\name{bi.format}
\alias{bi.format}
\title{bi.format}
\usage{
bi.format(filename, filetype = 0)
}
\arguments{
\item{filename}{input file name}

\item{filetype}{0, 1 is bmat}
}
\description{
add number of vertices and edges to the input file. The original input file will be changed
}
\examples{
bi.format(filename)

}

