% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R
\name{bi.print}
\alias{bi.print}
\title{Print the bicliques}
\usage{
bi.print(envir = .GlobalEnv$.bienv)
}
\arguments{
\item{envir}{biclique environment}
}
\description{
You can pass results from function bi.clique to this function to visualize the bicliques.
}
\examples{
bicliques = bi.clique(system.file("extdata", "example1.el", package = "biclique"))
bi.print()

}
