% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclique.R, R/package-biclique.R
\docType{package}
\name{biclique}
\alias{biclique}
\alias{biclique-package}
\title{biclique}
\usage{
biclique(filename, lleast = 1, rleast = 1, version = 1, getclique = 1,
  filetype = 0)
}
\arguments{
\item{filename}{input file name}

\item{lleast}{1}

\item{rleast}{1}

\item{version}{1}

\item{getclique}{1}

\item{filetype}{0}
}
\description{
Biclique enumeration

biclique package.
}
\examples{
bicliques = biclique("example1.el")

}
\author{
Yuping Lu
}
\keyword{biclique}

