% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-biclique.R
\docType{package}
\name{biclique}
\alias{biclique}
\alias{biclique-package}
\title{An R package to compute bicliques}
\description{
biclique package.
}
\author{
Yuping Lu, Charles Phillips, Michael Langston
}
\keyword{biclique}

