% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-biclique.R
\docType{package}
\name{biclique}
\alias{biclique}
\title{Maximal Biclique Enumeration in Bipartite Graphs}
\description{
A tool for enumerating maximal complete bipartite graphs. The input should be a edge list file or a binary matrix file. The output are maximal complete bipartite graphs. Algorithms used can be found in this paper Y. Lu et al. BMC Res Notes 13, 88 (2020) <doi:10.1186/s13104-020-04955-0>.
}
\details{
biclique package.
}
\author{
Yuping Lu, Charles Phillips, Michael Langston
}
\keyword{biclique}
